notification-services/pyproject.toml
[tool.poetry]
name = "notification-services"
version = "0.1.0"
description = ""
authors = ["MuhammadAhsaanAbbasi <mahsaanabbasi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.16"
pytest = "^8.0.2"
httpx = "^0.27.0"
itsdangerous = "^2.1.2"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240327"
python-multipart = "^0.0.9"
numpy = "^1.26.4"
types-requests = "^2.32.0.20240602"
protobuf = "^5.27.1"
aiokafka = "^0.10.0"
email-validator = "^2.1.1"
boto3 = "^1.34.143"
orjson = "^3.10.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
